requirements.txt
python-dotenv
pandas
numpy
matplotlib
scikit-learn
seaborn
pymongo[srv]
pymongo
certifi
dill
pyaml
mlflow
dagshub
fastapi
uvicorn

#-e .